+++ source/Relativity.Testing.Framework.Api/Strategies/Imaging/Environment/IGetMassImagingJobSizeStrategy.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+
+namespace Relativity.Testing.Framework.Api.Strategies
+{
+	internal interface IGetMassImagingJobSizeStrategy
+	{
+		int Get();
+
+		Task<int> GetAsync();
+	}
+}